Corrigindo atributos
alterando retorno de metódo de int para long

+++ app/src/main/java/com/example/refugiohotelsapp/Entidades/Reserva.java
@@ -115,7 +115,7 @@ import java.util.Date;
         this.reservaAtiva = reservaAtiva;
     }
 
-    public int getIdUsuario() {
+    public Long getIdUsuario() {
         return usuario.getIdUsuario();
     }
